Replaced enum fields with byte fields in all packets.

+++ atlas/src/test/java/us/ihmc/atlas/controllerAPI/AtlasUpperBodyTrajectoriesWhileWalkingTest.java
@@ -164,7 +164,7 @@ public class AtlasUpperBodyTrajectoriesWhileWalkingTest
             armTrajectoryMessage.setUniqueId(id);
             if (messageIndex > 0)
             {
-               armTrajectoryMessage.getQueueingProperties().setExecutionMode(ExecutionMode.QUEUE);
+               armTrajectoryMessage.getQueueingProperties().setExecutionMode(ExecutionMode.QUEUE.toByte());
                armTrajectoryMessage.getQueueingProperties().setPreviousMessageId(id - 1);
             }
             id++;
