+++ ihmc-avatar-interfaces/src/test/java/us/ihmc/avatar/controllerAPI/EndToEndArmTrajectoryMessageTest.java
@@ -278,7 +278,7 @@ public abstract class EndToEndArmTrajectoryMessageTest implements MultiRobotTest
          trajectoryMessage.setUniqueId(id);
          if (messageIndex > 0)
          {
-            trajectoryMessage.getQueueingProperties().setExecutionMode(ExecutionMode.QUEUE);
+            trajectoryMessage.getQueueingProperties().setExecutionMode(ExecutionMode.QUEUE.toByte());
             trajectoryMessage.getQueueingProperties().setPreviousMessageId(id - 1);
          }
          id++;
@@ -407,7 +407,7 @@ public abstract class EndToEndArmTrajectoryMessageTest implements MultiRobotTest
                if (messageIndex == numberOfMessages - 1)
                   previousMessageId = id + 100; // Bad ID
 
-               armTrajectoryMessage.getQueueingProperties().setExecutionMode(ExecutionMode.QUEUE);
+               armTrajectoryMessage.getQueueingProperties().setExecutionMode(ExecutionMode.QUEUE.toByte());
                armTrajectoryMessage.getQueueingProperties().setPreviousMessageId(previousMessageId);
             }
             id++;
@@ -513,7 +513,7 @@ public abstract class EndToEndArmTrajectoryMessageTest implements MultiRobotTest
             armTrajectoryMessage.setUniqueId(id);
             if (messageIndex > 0)
             {
-               armTrajectoryMessage.getQueueingProperties().setExecutionMode(ExecutionMode.QUEUE);
+               armTrajectoryMessage.getQueueingProperties().setExecutionMode(ExecutionMode.QUEUE.toByte());
                armTrajectoryMessage.getQueueingProperties().setPreviousMessageId(id - 1);
             }
             id++;
